Fixed a bug in util_sscanf_double. Was int tmp_value = strtof(. , .), but correct use is double tmp_value = strtof( . , .).
git-svn-id: http://ert.nr.no/repos/ert@1001 ecb61f28-d21e-0410-a4b6-a828cfe0ca3a

+++ devel/libutil/src/util.c
@@ -550,7 +550,7 @@ bool util_sscanf_double(const char * buffer , double * value) {
   bool value_OK = false; 
   char * error_ptr;
 
-  int tmp_value = strtod(buffer , &error_ptr);
+  double tmp_value = strtod(buffer , &error_ptr);
   if (error_ptr[0] == '\0') {
     value_OK = true; 
     *value = tmp_value;
